+++ fiddles/c/fiddle-0000-Template/template/main.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+int main(void) {
+    printf ("template compiled!");
+    return 0;
+}